+++ crates/producers/disolv-positions/src/vehicles/cache.rs
