+++ crates/producers/disolv-positions/src/vehicles/mod.rs
@@ -1,3 +1,3 @@
 pub(crate) mod cache;
-pub(crate) mod reader;
+pub(crate) mod trace;
 pub(crate) mod writer;
