+++ crates/producers/disolv-positions/src/vehicles/writer.rs
